extract view widget and adding notification screen
