+++ lib/utils/widget/bottom_navbar/custom_bottom_navbar.dart
@@ -4,11 +4,16 @@ import 'package:mobile_flutter/utils/widget/bottom_navbar/custom_navbar_provider
 import 'package:persistent_bottom_nav_bar_v2/persistent-tab-view.dart';
 import 'package:provider/provider.dart';
 
-class CustomBottomNavbar extends StatelessWidget {
+class CustomBottomNavbar extends StatefulWidget {
   const CustomBottomNavbar({
     super.key,
   });
 
+  @override
+  State<CustomBottomNavbar> createState() => _CustomBottomNavbarState();
+}
+
+class _CustomBottomNavbarState extends State<CustomBottomNavbar> {
   @override
   Widget build(BuildContext context) {
     return PersistentTabView(
@@ -29,19 +34,23 @@ class CustomBottomNavbar extends StatelessWidget {
         // borderRadius: BorderRadius.circular(10.0),
         colorBehindNavBar: Colors.white,
       ),
+
       popAllScreensOnTapOfSelectedTab: true,
       popActionScreens: PopActionScreensType.all,
+
       itemAnimationProperties: const ItemAnimationProperties(
         // Navigation Bar's items animation properties.
         duration: Duration(milliseconds: 200),
         curve: Curves.ease,
       ),
+
       screenTransitionAnimation: const ScreenTransitionAnimation(
         // Screen transition animation on change of selected tab.
         animateTabTransition: true,
         curve: Curves.ease,
         duration: Duration(milliseconds: 200),
       ),
+
       navBarStyle:
           NavBarStyle.simple, // Choose the nav bar style with this property.
     );
