+++ lib/view/home/screen/home_screen.dart
@@ -16,6 +16,7 @@ class _HomeScreenState extends State<HomeScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      resizeToAvoidBottomInset: false,
       body: Consumer<SharedPreferencesProvider>(
         builder: (context, sharedPrefs, _) {
           if (sharedPrefs.isLoading) {
